refactor: es_scripts -> ecospheres
