+++ setup.cfg
@@ -1,8 +1,8 @@
 [metadata]
-name = es_scripts
+name = ecospheres
 
 [options]
-packages = es_scripts
+packages = ecospheres
 
 [flake8]
 max-line-length=100
